A small server that for Computer Networks project
A small server that i have created for Computer Networks, It is basically a host app and a client app and multiple clients connect and can send messages, and the host can see the messages and their socket. It also numbers users in ascending order based on when they established a connection.